+++ src/admin/components/CommentsWidget/styles.js
