+++ src/admin/components/ReportManagement/index.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { adminApi } from '../../../../utils/api';
+import { adminApi } from '../../../utils/api';
 import {
   ManagementContainer,
   ManagementHeader,
