+++ src/user/components/SearchSection/styles.js
